Fashion-Recommendation-System/requirements.txt
tensorflow
numpy
scikit-learn
opencv-python
streamlit
Pillow
tqdm
fastapi
uvicorn
python-multipart
